fix review

+++ yatube/posts/views.py
@@ -126,17 +126,6 @@ def follow_index(request):
 @login_required
 def profile_follow(request, username):
     author = get_object_or_404(User, username=username)
-    is_subscribed = Follow.objects.filter(
-        user=request.user, author=author
-    ).exists()
-    if request.user != author and not is_subscribed:
-        Follow.objects.create(user=request.user, author=author)
-    return redirect('posts:profile', username=username)
-
-
-@login_required
-def profile_follow(request, username):
-    author = get_object_or_404(User, username=username)
     if request.user != author:
         Follow.objects.get_or_create(user=request.user, author=author)
     return redirect('posts:profile', username=username)
@@ -146,5 +135,6 @@ def profile_follow(request, username):
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
     follow_filter = Follow.objects.filter(user=request.user, author=author)
-    follow_filter.delete()
+    if follow_filter.exists():
+        follow_filter.delete()
     return redirect('posts:profile', username=username)
